Update README.md (#767)
I've updated the readme to include an important detail that will be helpful for newer developers or those less familiar with android documentation related to how to do the specific thing of filling AND cropping your splash screen:

android_gravity can be combined using the | operator to achieve multiple effects. For example:
`android_gravity: fill|clip_vertical` - This will fill the width while maintaining the image's vertical aspect ratio

This is one of the most basic and important things you need to get the splash working well on an Android tablet like an Amazon Fire.